added routine for not removing readme.md file from eddypro output folder

+++ ameriflux_pipeline/eddypro/runeddypro.py
@@ -45,14 +45,15 @@ class RunEddypro():
         # clean up the eddypro output folder
         print("All the contents in the", out_path, "will be removed.")
         for filename in os.listdir(out_path):
-            file_path = os.path.join(out_path, filename)
-            try:
-                if os.path.isfile(file_path) or os.path.islink(file_path):
-                    os.unlink(file_path)
-                elif os.path.isdir(file_path):
-                    shutil.rmtree(file_path)
-            except Exception as e:
-                print('Failed to delete %s. Reason: %s' % (file_path, e))
+            if filename.lower() != "readme.md":
+                file_path = os.path.join(out_path, filename)
+                try:
+                    if os.path.isfile(file_path) or os.path.islink(file_path):
+                        os.unlink(file_path)
+                    elif os.path.isdir(file_path):
+                        shutil.rmtree(file_path)
+                except Exception as e:
+                    print('Failed to delete %s. Reason: %s' % (file_path, e))
 
         # save temporary project file
         RunEddypro.save_string_list_to_file(tmp_proj_list, proj_file_name)
